Se agrego contenido al fragment de las recetas

+++ app/src/main/java/com/jonathan/retoyape/domain/model/Recipe.kt
@@ -8,8 +8,8 @@ data class Recipe(
     val name: String? = null,
     val description: String? = null,
     val locationName: String? = null,
-    val latitude: Long? = null,
-    val longitude: Long? = null,
+    val latitude: Double? = null,
+    val longitude: Double? = null,
 )
 
 fun Recipe.toRecipes() = RecipeEntity(
